backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==24.1.0
pydantic-settings==2.2.1
Pillow==10.4.0
pdf2image==1.17.0
python-docx==1.1.0
rich==13.7.1
httpx==0.28.1
